Added ability to select pre-registered land

+++ modules/General/TreeRemoval/Http/Controllers/TreeRemovalController.php
@@ -29,11 +29,19 @@ class TreeRemovalController extends Controller
 {
     public function openForm()
     {
+        $landbyuser = Process_Item::where([
+            ['status_id', '=', 5],
+            ['form_type_id', '=', 5],
+            ['created_by_user_id', '=', Auth()->user()->id],
+            ['request_organization', '!=', NULL],
+        ])->get()->pluck('form_id');
+        $landsdetails = Land_Parcel::whereIn('id', $landbyuser)->get();
         $gazettes = Gazette::all();
         $organizations = Organization::where('type_id', '<', '3')->get();
         return view('treeRemoval::form', [
             'organizations' => $organizations,
             'gazettes' => $gazettes,
+            'registered_lands' => $landsdetails,
         ]);
     }
 
@@ -48,83 +56,116 @@ class TreeRemovalController extends Controller
             ]);
         }
 
-        $request->validate([
-            'planNo' => 'required',
-            'surveyorName' => 'required',
-            'district' => 'required|exists:districts,district',
-            'gs_division' => 'nullable|exists:gs_divisions,gs_division',
-            'polygon' => 'required',
-            'number_of_trees' => 'required|integer',
-            'description' => 'required',
-            'land_extent' => 'nullable|numeric|between:0,99.999',
-            'number_of_tree_species' => 'nullable|integer',
-            'number_of_flora_species' => 'nullable|integer',
-            'number_of_reptile_species' => 'nullable|integer',
-            'number_of_mammal_species' => 'nullable|integer',
-            'number_of_amphibian_species' => 'nullable|integer',
-            'number_of_fish_species' => 'nullable|integer',
-            'number_of_avian_species' => 'nullable|integer',
-            'externalRequestor' => 'nullable|regex:/^[0-9]{9}[vVxX]$/',
-            'erEmail' => 'nullable|email',
-            'land_gazettes' => 'nullable',
-            'land_governing_orgs' => 'nullable',
-            'location.*.tree_species_id' => 'required',
-            'location.*.circumference_at_breast_height' => 'required|numeric|between:0,999.999',
-            'location.*.height'    => 'required||numeric|between:0,999.999',
-        ]);
+        if (request('registered_land')) {
+            $request->validate([
+                'planNo' => 'nullable',
+                'surveyorName' => 'nullable',
+                'district' => 'required|exists:districts,district',
+                'gs_division' => 'nullable|exists:gs_divisions,gs_division',
+                'polygon' => 'nullable',
+                'land_extent' => 'nullable|numeric|between:0,99.999',
+                'land_gazettes' => 'nullable',
+                'land_governing_orgs' => 'nullable',
+                'registered_land' => 'required|exists:land_parcels,id',
+
+                'number_of_trees' => 'required|integer',
+                'description' => 'required',
+                'number_of_tree_species' => 'nullable|integer',
+                'number_of_flora_species' => 'nullable|integer',
+                'number_of_reptile_species' => 'nullable|integer',
+                'number_of_mammal_species' => 'nullable|integer',
+                'number_of_amphibian_species' => 'nullable|integer',
+                'number_of_fish_species' => 'nullable|integer',
+                'number_of_avian_species' => 'nullable|integer',
+                'externalRequestor' => 'nullable|regex:/^[0-9]{9}[vVxX]$/',
+                'erEmail' => 'nullable|email',
+                'location.*.tree_species_id' => 'required',
+                'location.*.circumference_at_breast_height' => 'required|numeric|between:0,999.999',
+                'location.*.height'    => 'required||numeric|between:0,999.999',
+            ]);
+        } else {
+            $request->validate([
+                'planNo' => 'required',
+                'surveyorName' => 'required',
+                'district' => 'required|exists:districts,district',
+                'gs_division' => 'nullable|exists:gs_divisions,gs_division',
+                'polygon' => 'required',
+                'number_of_trees' => 'required|integer',
+                'description' => 'required',
+                'land_extent' => 'nullable|numeric|between:0,99.999',
+                'number_of_tree_species' => 'nullable|integer',
+                'number_of_flora_species' => 'nullable|integer',
+                'number_of_reptile_species' => 'nullable|integer',
+                'number_of_mammal_species' => 'nullable|integer',
+                'number_of_amphibian_species' => 'nullable|integer',
+                'number_of_fish_species' => 'nullable|integer',
+                'number_of_avian_species' => 'nullable|integer',
+                'externalRequestor' => 'nullable|regex:/^[0-9]{9}[vVxX]$/',
+                'erEmail' => 'nullable|email',
+                'land_gazettes' => 'nullable',
+                'land_governing_orgs' => 'nullable',
+                'location.*.tree_species_id' => 'required',
+                'location.*.circumference_at_breast_height' => 'required|numeric|between:0,999.999',
+                'location.*.height'    => 'required||numeric|between:0,999.999',
+            ]);
+        }
 
         DB::transaction(function () use ($request) {
-
-            $land = new Land_Parcel();
-            $land->title = request('planNo');
-            $land->surveyor_name = request('surveyorName');
-
-            $land->polygon = request('polygon');
-
-            $land->created_by_user_id = request('createdBy');
-
-            if (request('isProtected')) {
-                $land->protected_area = request('isProtected');
-            }
+            //used by a function outside the for loop below
+            $gs_division_id1 = GS_Division::where('gs_division', request('gs_division'))->pluck('id');
             $district_id1 = District::where('district', request('district'))->pluck('id');
-            $land->district_id = $district_id1[0];
+            $org_id = $request->organization;
 
-            if (request('gs_division')) {
-                $gs_division_id1 = GS_Division::where('gs_division', request('gs_division'))->pluck('id');
-                $land->gs_division_id = $gs_division_id1[0];
-            }
+            if (!(request('registered_land'))) {
+                $land = new Land_Parcel();
+                $land->title = request('planNo');
+                $land->surveyor_name = request('surveyorName');
 
+                $land->polygon = request('polygon');
 
-            if (($request->organization) != null) {
-                $org_id = request('organization');
-                $land->activity_organization = $org_id;
-            }
-            $land->status_id = 1;
-            $land->save();
+                $land->created_by_user_id = request('createdBy');
 
-            $landid = Land_Parcel::latest()->first()->id;
+                if (request('isProtected')) {
+                    $land->protected_area = request('isProtected');
+                }
 
-            if (request('land_governing_orgs')) {
-                $governing_organizations = request('land_governing_orgs');
+                $land->district_id = $district_id1[0];
 
-                foreach ($governing_organizations as $governing_organization) {
-                    $land_has_organization = new Land_Has_Organization();
-                    $land_has_organization->land_parcel_id = $landid;
-                    $land_has_organization->organization_id = $governing_organization;
-                    $land_has_organization->save();
+                if (request('gs_division')) {
+                    $land->gs_division_id = $gs_division_id1[0];
+                }
+
+                if (($request->organization) != null) {
+                    $land->activity_organization = $org_id;
+                }
+                $land->status_id = 1;
+                $land->save();
+
+                $landid = Land_Parcel::latest()->first()->id;
+
+                if (request('land_governing_orgs')) {
+                    $governing_organizations = request('land_governing_orgs');
+
+                    foreach ($governing_organizations as $governing_organization) {
+                        $land_has_organization = new Land_Has_Organization();
+                        $land_has_organization->land_parcel_id = $landid;
+                        $land_has_organization->organization_id = $governing_organization;
+                        $land_has_organization->save();
+                    }
+                }
+
+                if (request('land_gazettes')) {
+                    $gazettes = request('land_gazettes');
+
+                    foreach ($gazettes as $gazette) {
+                        $land_has_gazette = new Land_Has_Gazette();
+                        $land_has_gazette->land_parcel_id = $landid;
+                        $land_has_gazette->gazette_id = $gazette;
+                        $land_has_gazette->save();
+                    }
                 }
             }
 
-            if (request('land_gazettes')) {
-                $gazettes = request('land_gazettes');
-
-                foreach ($gazettes as $gazette) {
-                    $land_has_gazette = new Land_Has_Gazette();
-                    $land_has_gazette->land_parcel_id = $landid;
-                    $land_has_gazette->gazette_id = $gazette;
-                    $land_has_gazette->save();
-                }
-            }
 
             $tree = new Tree_Removal_Request();
             //Required and/or filled in fields
@@ -170,13 +211,17 @@ class TreeRemovalController extends Controller
 
             $tree->status_id = 1;
 
-            $tree->land_parcel_id = $landid;
+            if (request('registered_land')) {
+                $tree->land_parcel_id = request('registered_land');
+            } else {
+                $tree->land_parcel_id = $landid;
+            }
 
             //calculating volume
             $locations = request('location');
             foreach ($locations as &$location) {
-                $radius = $location['circumference_at_breast_height']/(2*pi());
-                $volume = pi()* $radius * $radius * $location['height'];
+                $radius = $location['circumference_at_breast_height'] / (2 * pi());
+                $volume = pi() * $radius * $radius * $location['height'];
                 $location['timber_volume'] = $volume;
             }
             $tree->tree_details = $locations;
@@ -209,7 +254,7 @@ class TreeRemovalController extends Controller
             } else {
                 $treeProcess->request_organization = auth()->user()->organization_id;
             }
-            if (($request->activity_organization) != null) {
+            if (($request->organization) != null) {
                 $treeProcess->activity_organization = $org_id;
             }
 
@@ -219,32 +264,37 @@ class TreeRemovalController extends Controller
 
             $latestTreeProcess = Process_Item::latest()->first();
 
-            if (($request->activity_organization) == null) {
+            if (($request->organization) == null) {
                 $org_id = organization_assign::auto_assign($latestTreeProcess->id, $district_id1[0]);
-                Land_Parcel::where('id', $landid)->update(['activity_organization' => $org_id]);
+                if (!(request('registered_land'))) {
+                    Land_Parcel::where('id', $landid)->update(['activity_organization' => $org_id]);
+                }
                 Tree_Removal_Request::where('id', $latest->id)->update(['organization_id' => $org_id]);
-            }else{
+            } else {
                 $users = User::where('role_id', '<', 3)->get();
                 Notification::send($users, new ApplicationMade($latestTreeProcess));
             }
-            $landProcess = new Process_Item();
-            $landProcess->form_id = $landid;
-            $landProcess->remark = "Verify these land details";
-            $landProcess->prerequisite = 0;
+            if (!(request('registered_land'))) {
+                $landProcess = new Process_Item();
+                $landProcess->form_id = $landid;
+                $landProcess->remark = "Verify these land details";
+                $landProcess->prerequisite = 0;
 
-            if (request('checkExternalRequestor')) {
-                $landProcess->ext_requestor = request('externalRequestor');
-                $landProcess->ext_requestor_email = request('erEmail');
-            } else {
-                $landProcess->request_organization = auth()->user()->organization_id;
+                if (request('checkExternalRequestor')) {
+                    $landProcess->ext_requestor = request('externalRequestor');
+                    $landProcess->ext_requestor_email = request('erEmail');
+                } else {
+                    $landProcess->request_organization = auth()->user()->organization_id;
+                }
+                $landProcess->activity_organization = $org_id;
+
+                $landProcess->status_id = 1;
+                $landProcess->form_type_id = 5;
+                $landProcess->created_by_user_id = request('createdBy');
+                $landProcess->prerequisite_id = $latestTreeProcess->id;
+                $landProcess->save();
             }
-            $landProcess->activity_organization = $org_id;
 
-            $landProcess->status_id = 1;
-            $landProcess->form_type_id = 5;
-            $landProcess->created_by_user_id = request('createdBy');
-            $landProcess->prerequisite_id = $latestTreeProcess->id;
-            $landProcess->save();
 
             //making a downloadable version of the KML file
             if (request('kml') !== null) {  //if the file is uploaded then the kml file will not be created
@@ -288,27 +338,29 @@ class TreeRemovalController extends Controller
 
         DB::transaction(function () use ($processid, $treeid, $landid, $prereqs) {
 
-            $landhasGazettes = Land_Has_Gazette::where("land_parcel_id", "=", $landid)->get();
-            foreach ($landhasGazettes as $landhasGazette) {
-                $landhasGazette->delete();
-            }
+            if (!($prereqs->isEmpty())) {
+                $landhasGazettes = Land_Has_Gazette::where("land_parcel_id", "=", $landid)->get();
+                foreach ($landhasGazettes as $landhasGazette) {
+                    $landhasGazette->delete();
+                }
 
-            $landHasOrganizations = Land_Has_Organization::where("land_parcel_id", "=", $landid)->get();
-            foreach ($landHasOrganizations as $landHasOrganization) {
-                $landHasOrganization->delete();
-            }
+                $landHasOrganizations = Land_Has_Organization::where("land_parcel_id", "=", $landid)->get();
+                foreach ($landHasOrganizations as $landHasOrganization) {
+                    $landHasOrganization->delete();
+                }
 
-            $landParcelProcess = Process_Item::find($prereqs[0]);
-            $landParcelProcess->delete();
+                $landParcelProcess = Process_Item::find($prereqs[0]);
+                $landParcelProcess->delete();
+
+                $landParcel = Land_Parcel::find($landid);
+                $landParcel->delete();
+            }
 
             $treeRemovalProcess = Process_Item::find($processid);
             $treeRemovalProcess->delete();
 
             $treeRemoval = Tree_Removal_Request::find($treeid);
             $treeRemoval->delete();
-
-            $landParcel = Land_Parcel::find($landid);
-            $landParcel->delete();
         });
         return redirect('/approval-item/showRequests')->with('message', 'Request Successfully Deleted');
     }
